Split up signs and diagnostics
